app/requirements.txt
flask==3.0.3
flask-socketio==5.4.1
flasgger==0.9.7b2
eventlet==0.38.0
sqlalchemy-orm==1.2.10
sqlalchemy==2.0.36
psycopg==3.2.3
psycopg-binary==3.2.3
toml==0.10.2
requests==2.25.1
pydantic==2.9.2
python-dotenv==1.0.1